Floats (#172)
* switch to returning a matrix of float values instead of doubles

* change to floats

* update git ignore

* update vector float

* replace double with float, have to change some test cases since the precision changes

+++ src/Algorithms/Algorithm.hpp
@@ -20,7 +20,7 @@ typedef unsigned int algorithm_id_t;
 
 class Algorithm {
 protected:
-    double progress;
+    float progress;
     int maxIteration;
     bool isRunning;
     bool shouldStop;
@@ -33,7 +33,7 @@ public:
 
     void setMaxIteration(int);
 
-    double getProgress();
+    float getProgress();
     bool getIsRunning();
     int getMaxIteration();
     void stop();
